@@ -4,25 +4,23 @@
   "private": true,
   "dependencies": {
     "@ant-design/plots": "^2.3.2",
-    "@ant-design/pro-components": "^2.6.2",
+    "@ant-design/pro-components": "^2.8.2",
     "@craco/craco": "^7.1.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.121",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-router": "^5.1.20",
-    "@types/react-router-dom": "^5.3.3",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "antd": "^5.22.3",
     "axios": "^1.7.8",
     "classnames": "^2.5.1",
     "dayjs": "^1.11.13",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router": "^7.0.2",
+    "react-router-dom": "^7.0.2",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
